Fix VTI pressure measurements after vti_ps_stop
When vti_ps_stop has been called and the sensor is in standy, it seems
that data for the last measurement may still arrive. If that data is
not read, it seems to block further measurement after entering ultra
low power mode again.

To work around that, read pending data in vti_ps_start and discard it.