Change PS interrupt handling
The interrupt handlers store results and users query the stored
results instead of doing the result reading inline.

This is a prepartion for the requirement of the BMP sensor,
which requires the temperature to be read in order to calculate
the pressure result. This requires two measurement (and interrupts)
to be done in sequence.