First working BMP sensor driver
For getting the correct pressure value the temperature sensor has to
be queried first as one of the temperature coefficients is needed for
calculating the pressure.

Querying the sensor is decoupled now from providing the value to users.
The ps message event is generated when one temperature+pressure cycle
has been completed.