Switch pressure sensor to time-based measurements
Pressure measurements are being started by the RTC_SECOND event now.
This will make it easier to modify the measurement interval to reduce
power consumption.

There are still some oddities: during the time when backlight is
enabled or when buttons are pressed it appears as if the measurement
interval is reduced. This is visible by looking at the up/down markers
in altitude mode, which change for every pressure event.